move guard definition
