+++ frontend/src/app/routes.ts
@@ -7,7 +7,7 @@ import { RegisterComponent } from './user/register/register.component';
 import { PasswordForgottenComponent } from './user/password-forgotten/password-forgotten.component';
 import { ResetPasswordComponent } from './user/reset-password/reset-password.component';
 import { UserSettingsComponent } from './user/user-settings/user-settings.component';
-import { ContributorGuard, LoggedOnGuard } from '@shared/logged-on.guard';
+import { LoggedOnGuard } from '@shared/logged-on.guard';
 import { SourcesComponent } from './data-entry/sources/sources.component';
 import { LocationFormComponent } from './data-entry/location-form/location-form.component';
 import { GiftFormComponent } from './data-entry/gift-form/gift-form.component';
@@ -19,6 +19,7 @@ import {
     SITE_NAME, sourceFormTitleResolver, spaceFormTitleResolver
 } from './titles';
 import { EpisodeFormComponent } from './data-entry/episode-form/episode-form.component';
+import { ContributorGuard } from '@shared/contributor.guard';
 
 
 const routes: Routes = [
