+++ frontend/src/app/shared/logged-on.guard.ts
@@ -26,27 +26,3 @@ export const LoggedOnGuard: CanActivateFn = (route: ActivatedRouteSnapshot) => {
         })
     );
 };
-
-export const ContributorGuard: CanActivateFn = (route: ActivatedRouteSnapshot) => {
-    const authService = inject(AuthService);
-    const toastService = inject(ToastService);
-    const router = inject(Router);
-
-    return authService.currentUser$.pipe(
-        filter((user) => user !== undefined),
-        map((user) => {
-            if (user?.isContributor) {
-                return true;
-            } else {
-                let body = 'You do not have permission to view this page.';
-                if (!user) { body += ' Do you need to sign in?'; }
-                toastService.show({
-                    type: 'danger',
-                    header: 'Not authorised',
-                    body,
-                });
-                return router.createUrlTree(['/']);
-            }
-        }),
-    );
-};
